requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pinecone-client>=5.0.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pdfplumber>=0.11.4
openai==1.76.0
firebase-admin==6.8.0
celery==5.3.4
redis==5.0.1
